runTests and runTestsContinuously accept arguments
Instead of hardcoding the arguments in runTests, let's allow the user to
choose them. Of course, this means that the user could forget to specify
them or specify them incorrectly, so now we need some logic to test the
arguments. We already had some logic to do so, but the error message was
duplicated on all the nodes, resulting in an interleaved mess. So I
added some new logic to test the arguments once, before launching the
nodes.