network-transport-tcp's hello world
I can't rely on distributed-process-simplelocalnet because it is using a
multicast to discover the other nodes, this won't work if the nodes are
in different subnets. Also, we already know the list of nodes from
nodelist.txt, so we don't need the multicast in the first place.

The TCP example uses a known host and a port to get two processes to
communicate, maybe I can adapt this to multiple interacting nodes?