run_tests.sh can now be used by the user
It was previously intended to be used as a helper for
run_tests_continuously.sh, but that role has been moved to
scripts/run_tests_and_hang.sh so that the name run_tests.sh can be used
for a script which runs the tests once.

Note that we currently have to terminate the test manually using CTRL-C,
as before, but at least now doing so will clean up the nodes which were
spawned during the test.