from a list to maybes and back
Instead of directly returning the list of messages which receiveMany is
returning, we now store this list and return one message at a time. The
goal is not to make it looks like we are receiving one message at a
time: on the contrary, receiveContributions is recovering the original
list. To do so, instead of returning one message at a time, we return a
Maybe message, and Nothing is a sentinel indicating the end of the list.

The reason we go through those hoops is to make it easier for the
interpreter to treat some individual messages specially without having
to search through a list of messages and to reconstruct the list
afterwards.