demonstrate that we're within the grace limit
We were within the grace limit before too, I was just looking at the
wrong performance number.

    real    0m11.358s
    user    0m12.894s
    sys     0m10.875s

I assumed that the largest number had to be the wall clock time, but it
turns out that when more than one CPU is involved, the "user" time can
be greated than the "real" wall clock time.