receive contributions in a separate thread
I tried to run the program on two machines connected via wifi, and I was
pleasantly surprised to discover that the connection between the nodes
was automatically restored after temporarily disabling the wifi on one
of the machines, despite the fact that I'm not handling disconnection
errors yet.

I then tried to keep the wifi disabled until the end of the grace
period, and I was disappointed not to see any output. The reason is that
self-connection which my helper thread is using to communicate with the
main thread is based on the IP which the wifi router is assigning the
node, so when the wifi is down, internal communication stops as well.

The solution is to communicate via an MVar instead. Since this means
that the main thread could now be blocked either on an MVar or on
receiveMany, I added another helper thread whose sole task is to
repeatedly call receiveMany and to forward the result to the main thread
via the MVar, this way the main thread only ever needs to block on the
MVar.

As an added benefit, this new design ensures that the only kind of
messages which receiveMany ever receives are contributions, eliminating
the need for the complicated scheme in which I divided the list of
contributions into single elements and later reconstructed the list.