avoid "inconsistent previous messages" error
When a message indicates that the other node has already started a new
round, we must accept its previousMessage field as the result of the
current round, and then contribute to the new round. When contributing
to the new round, I basing my new contribution on the contribution I had
about the previous round, but that value was incorrect because it didn't
contain the final message for the previous round. This error was
detected by an inconsistency check.